+++ selfdrive/car/chrysler/carcontroller.py
@@ -116,7 +116,7 @@ class CarController(CarControllerBase):
 
       self.apply_steer_last = apply_steer
 
-      can_sends.append(chryslercan.create_lkas_command(self.packer, self.CP, int(apply_steer), lkas_control_bit, self.steerNoMinimum, CC.latActive))
+      can_sends.append(chryslercan.create_lkas_command(self.packer, self.CP, int(apply_steer), lkas_control_bit, CC.latActive))
 
     if CC.enabled:
       # auto set profile